Colormap optimization supports RS bag file
