+++ examples/python/utility/file.py
@@ -5,9 +5,11 @@
 # examples/python/utility/file.py
 
 from os import listdir, makedirs
-from os.path import exists, isfile, join, splitext
+from os.path import exists, isfile, join, splitext, dirname, basename
 import shutil
 import re
+from warnings import warn
+import json
 import open3d as o3d
 
 
@@ -105,3 +107,26 @@ def read_poses_from_log(traj_log):
             trans_arr.append(T_gt)
 
     return trans_arr
+
+
+def extract_rgbd_frames(rgbd_video_file):
+    """
+    Extract color and aligned depth frames and intrinsic calibration from an
+    RGBD video file (currently only RealSense bag files supported). Folder
+    structure is:
+        <directory of rgbd_video_file/<rgbd_video_file name without extension>/
+            {depth/00000.jpg,color/00000.png,intrinsic.json}
+    """
+    frames_folder = join(dirname(rgbd_video_file),
+                         basename(splitext(rgbd_video_file)[0]))
+    path_intrinsic = join(frames_folder, "intrinsic.json")
+    if isfile(path_intrinsic):
+        warn(f"Skipping frame extraction for {rgbd_video_file} since files are"
+             " present.")
+    else:
+        rgbd_video = o3d.t.io.RGBDVideoReader.create(rgbd_video_file)
+        rgbd_video.save_frames(frames_folder)
+    with open(path_intrinsic) as intr_file:
+        intr = json.load(intr_file)
+    depth_scale = intr["depth_scale"]
+    return frames_folder, path_intrinsic, depth_scale
